Add a progress bar.
There's unfortunately no Drive API to get the total amount of files
in a drive. Use the total storage as a proxy.

Two problems with this:
1. Progress will be less linear as some files might account for lots of
storage, others less - so different `files.list()` responses will result
in widely different progress jumps. The bar might be stuck on 5% for an
hour, then jump to 100% in a second.
2. The sum of all file sizes does not necessarily match the total
storage reported by Drive.  For instance, files with multiple parents
will be counted multiple times in our case.  Space taken by "Google
Photos/" in the root doesn't seem accounted in my case.  Docs/sheets may
be treated specially.  We might complete the Drive walk before or after
we reach 100%.

We revisit to cover these problems better.  For now a naive solution to
get going.