Reimplement zoom.
The current zoom is implemented using a svg transform, which assumes a
fully  populated SVG.  To save on memory, we will start pruning the SVG
tree, and even the d3 hierarchy.  We'll want the user to zoom into areas
to see details. The new approach reduces the d3 hierarchy to only the
zoomed in node and its descendants.  Zooming out goes up one level.