Take 2 on previous commit, preserving colors on zoom/unzoom.
Colors are determined from the d3 hierarchy.  The previous commit
implements zoom/unzoom by preprocessing the input to the d3 hierarchy.
Colors are therefor not stable.  It is easy to change the color
computation to be based on the preprocessed data, except it won't apply
to the same order of children, as reordering occurs only in the d3
hierarchy (and is also based on the sum, which requires a different data
strucutre).

So, a different way to do this is to prune the tree to the node zoomed
in once we're dealing with the d3 hierarchy.  Luckily d3 provides a copy
operator that returns the node as a new tree root.