Implement a max depth.
Why? Prune details that adds to computational complexity and are
unlikely to be useful in a single view.  The user can zoom into lower
nodes to dig deeper.

Why a max depth of 2 (root + 2 levels)?  This will be upped in the
future, just playing around with a restrictive setting for now.