Add a table with child sizes.
This has the basic interaction we want.  Styling isn't quite there yet.
The table has too much space between rows, scrolling to the right is
sometimes enabled, numbers and units get split over different lines,
etc.  There should also still be a title with the name of the parent of
the table children, and that total size.