Color rects according to their relative X center.
Why?  There are a couple of issues with the current color scheme:
1. Only root children colors vary. The root node is a separate (fixed)
color, and descendents of root children use their parent in the root
children as color.
1. On the icicle graph, the colors don't appear to vary enough. Because
the number of children is used to split the range, and the icicle graph
only shows like the top 1% - others are too small, in the case of My
Drive (representative?), only a small portion of the start of the color
range is used.

This new way to map nodes to colors solves both. Unfortunately, it
introduces new problems:
1. The many small directories now all look indistinguishable in terms
of color.  That seems ok in the icicle graph, b/c those rects are also
very small, but looks ugly in the table, where the long tail of small
dirs appear to all be the same color.  The previous scheme had exactly
what we wanted for the table in terms of color variations.
1. The icicles no longer are distinguishable. Because all that matters
is the relative X, not sharing some ancestry.

This doesn't appear to be a clear net win, but I am submitting it
nevertheless as either scheme will need to be revisited, and this way I
document the alternatives. I will ponder about another attempt, likely
combining these approaches somewhat.  (Maybe go back to children size
based range split, and attribute a color range close to a parent's color
to keep the icicles standing out.)