Allow initial undefined Drive data.
Why? It is an expected app state. An empty My Drive is not the same.
Handle by redirecting to the main page.

For now, move the start button to the main page.  We might still want
buttons to restart and logout on the Drive page, but unsure yet.
Clicking on the logo to go back to the start page, or letting the user
browser navigate there might be sufficient.